# deploy/nginx/html/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Verifi Agent API",
    "summary": "Human judgment your AI agent can call, paid per request with x402.",
...
    "x-guidance": "POST /verify with intent, claim and your wallet as agent_id. Pay the 402: it lists USDC first and EURC second and carries every later price in extensions.service-windows. Store verify_id, then wait for the callback or poll GET /verify/{verify_id} (free). When status is ready, POST /verify-unlock?id={verify_id} and pay its single requirement, 2.90 EUR if the human answered within 60 minutes of admission, 1.45 EUR if later, in the asset of the admission. Read the result from the completed response. Never send a private key.",
    "version": "3.0.0",
    "contact": {
      "name": "Verifi",
      "url": "https://verifi.cloud/contact"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://verifi.cloud/"
    }
  },
  "servers": [
    {
      "url": "https://verifi.cloud",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/verify/{verify_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/VerifyIdPath"
        }
      ],
      "get": {
        "tags": [
          "verify"
        ],
        "operationId": "getVerify",
        "summary": "Poll the state of a chain",
        "description": "Returns the current state. Polling never costs money and never reveals a locked result.\n\nWhile status is processing, honor retry_after_seconds. A processing response carries admitted_at, sla_deadline, and grace_deadline, so the agent can plan its polling; expires_at equals grace_deadline, the last deadline. Human work can take minutes or hours, up to 24 hours, so do not assume a short fixed deadline. Prefer callback_url and use polling as the recovery path.\n\nA ready response states exactly what the unlock costs and why: service_window.applied (sla or grace, decided by ready_at alone) and service_window.unlock_amount in the bound asset.",
        "responses": {
          "200": {
            "description": "Current state of the chain.",
            "headers": {
              "Retry-After": {
                "$ref": "#/components/headers/RetryAfter"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyState"
                },
                "examples": {
                  "processing": {
                    "summary": "A human is working. Poll again.",
                    "value": {
                      "verify_id": "0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11",
                      "work_id": "0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11",
                      "contract_version": 3,
                      "status": "processing",
                      "verdict": null,
                      "explanation": null,
                      "response": null,
                      "admitted_at": "2026-10-01T09:00:02+00:00",
                      "sla_deadline": "2026-10-01T10:00:02+00:00",
                      "grace_deadline": "2026-10-02T09:00:02+00:00",
                      "ready_at": null,
                      "expires_at": "2026-10-02T09:00:02+00:00",
                      "next_action": "poll",
                      "poll_url": "/verify/0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11",
                      "retry_after_seconds": 15
                    }
                  },
                  "readySlaUsdc": {
                    "summary": "USDC chain. The human answered inside the SLA window, so the unlock costs the SLA amount.",
                    "value": {
                      "verify_id": "0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11",
                      "work_id": "0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11",
                      "contract_version": 3,
                      "status": "ready",
                      "verdict": null,
                      "explanation": null,
                      "response": null,
                      "admitted_at": "2026-10-01T09:00:02+00:00",
                      "sla_deadline": "2026-10-01T10:00:02+00:00",
                      "grace_deadline": "2026-10-02T09:00:02+00:00",
                      "ready_at": "2026-10-01T09:41:15+00:00",
                      "responded_at": "2026-10-01T09:41:15+00:00",
                      "next_action": "unlock",
                      "unlock_url": "https://verifi.cloud/verify-unlock?id=0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11",
                      "unlock": {
                        "method": "POST",
                        "url": "/verify-unlock?id=0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11",
                        "price_usdc": "3.40",
                        "payment_required": true,
                        "funded_by": "x402",
                        "network": "eip155:8453",
                        "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                        "asset_symbol": "USDC",
                        "amount_atomic": "3400000",
                        "amount_decimal": "3.40",
                        "applied_window": "sla"
                      },
                      "service_window": {
                        "applied": "sla",
                        "sla_deadline": "2026-10-01T10:00:02+00:00",
                        "grace_deadline": "2026-10-02T09:00:02+00:00",
                        "network": "eip155:8453",
                        "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                        "unlock_amount": "3400000"
                      }
                    }
                  },
                  "readyGraceEurc": {
                    "summary": "EURC chain. The human answered after the SLA window, so the unlock costs the grace amount.",
                    "value": {
                      "verify_id": "6b2a1f6e-9a0d-4a2b-91c6-8f0f2d5a7c33",
                      "work_id": "6b2a1f6e-9a0d-4a2b-91c6-8f0f2d5a7c33",
                      "contract_version": 3,
                      "status": "ready",
                      "verdict": null,
                      "explanation": null,
                      "response": null,
                      "admitted_at": "2026-10-01T09:00:02+00:00",
                      "sla_deadline": "2026-10-01T10:00:02+00:00",
                      "grace_deadline": "2026-10-02T09:00:02+00:00",
                      "ready_at": "2026-10-01T13:12:40+00:00",
                      "responded_at": "2026-10-01T13:12:40+00:00",
                      "next_action": "unlock",
                      "unlock_url": "https://verifi.cloud/verify-unlock?id=6b2a1f6e-9a0d-4a2b-91c6-8f0f2d5a7c33",
                      "unlock": {
                        "method": "POST",
                        "url": "/verify-unlock?id=6b2a1f6e-9a0d-4a2b-91c6-8f0f2d5a7c33",
                        "price_usdc": null,
                        "payment_required": true,
                        "funded_by": "x402",
                        "network": "eip155:8453",
                        "asset": "0x60a3E35Cc302bFA44Cb288Bc5a4F316Fdb1adb42",
                        "asset_symbol": "EURC",
                        "amount_atomic": "1450000",
                        "amount_decimal": "1.45",
                        "applied_window": "grace"
                      },
                      "service_window": {
                        "applied": "grace",
                        "sla_deadline": "2026-10-01T10:00:02+00:00",
                        "grace_deadline": "2026-10-02T09:00:02+00:00",
                        "network": "eip155:8453",
                        "asset": "0x60a3E35Cc302bFA44Cb288Bc5a4F316Fdb1adb42",
                        "unlock_amount": "1450000"
                      }
                    }
                  },
                  "failedEurc": {
                    "summary": "Nobody answered within 24 hours. Stop. The next admission of this wallet is free, in EURC.",
                    "value": {
                      "verify_id": "6b2a1f6e-9a0d-4a2b-91c6-8f0f2d5a7c33",
                      "work_id": "6b2a1f6e-9a0d-4a2b-91c6-8f0f2d5a7c33",
                      "contract_version": 3,
                      "status": "failed",
                      "next_action": "stop",
                      "failure": {
                        "reason": "human_timeout",
                        "entry_credit_granted": true,
                        "entry_credit_value_usdc": null,
                        "entry_credit": "next_admission"
                      }
                    }
                  },
                  "failedUsdc": {
                    "summary": "The same on a USDC chain: entry_credit_value_usdc is the admission price paid.",
                    "value": {
                      "verify_id": "0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11",
                      "work_id": "0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11",
                      "contract_version": 3,
                      "status": "failed",
                      "next_action": "stop",
                      "failure": {
                        "reason": "human_timeout",
                        "entry_credit_granted": true,
                        "entry_credit_value_usdc": "0.12",
                        "entry_credit": "next_admission"
                      }
                    }
                  },
                  "completed": {
                    "summary": "The unlock already passed. The answer is visible.",
                    "value": {
                      "verify_id": "6b2a1f6e-9a0d-4a2b-91c6-8f0f2d5a7c33",
                      "work_id": "6b2a1f6e-9a0d-4a2b-91c6-8f0f2d5a7c33",
                      "contract_version": 3,
                      "status": "completed",
                      "human_status": "refined",
                      "verdict": "refined",
                      "explanation": "Correct except the date. Pricing starts 1 October, not 1 September.",
                      "response": "Correct except the date. Pricing starts 1 October, not 1 September.",
                      "response_time_ms": 15278000,
                      "ready_at": "2026-10-01T13:12:40+00:00",
                      "unlocked_at": "2026-10-01T13:20:05+00:00",
                      "next_action": "done"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The verify id is not a UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "invalid verify id"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/BackendUnavailable"
          }
        },
        "security": []
      }
    },
...
    "/verify/{verify_id}/unlock": {
      "parameters": [
        {
          "$ref": "#/components/parameters/VerifyIdPath"
        }
      ],
      "post": {
        "tags": [
          "verify"
        ],
        "operationId": "unlockVerifyLegacy",
        "summary": "Legacy unlock path",
        "description": "Permanent redirect to POST /verify-unlock?id={verify_id}. The method and body are preserved.",
        "responses": {
          "308": {
            "description": "Permanent redirect to the canonical unlock endpoint.",
            "headers": {
              "Location": {
                "description": "The canonical unlock URL.",
                "schema": {
                  "type": "string",
                  "examples": [
                    "/verify-unlock?id=0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "VerifyIdPath": {
        "name": "verify_id",
        "in": "path",
        "required": true,
        "description": "The durable chain handle returned by POST /verify.",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "example": "0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11"
      },
...
    },
    "headers": {
      "RetryAfter": {
        "description": "Seconds to wait before the next call.",
        "schema": {
          "type": "integer"
        },
        "example": 15
      },
...
    },
    "responses": {
...
      "NotFound": {
        "description": "No chain exists for that verify_id.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "verify not found"
            }
          }
        }
      },
      "BackendUnavailable": {
        "description": "The verification backend is temporarily unavailable. Nothing was charged. Retry.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "verification backend unavailable"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ServiceWindow": {
        "type": "object",
        "description": "Why the unlock costs what it costs. Present when status is ready on a contract 3 chain.",
        "properties": {
          "applied": {
            "type": "string",
            "enum": [
              "sla",
              "grace"
            ],
            "description": "sla if ready_at is at or before sla_deadline, otherwise grace. A boundary instant belongs to the earlier window."
          },
          "sla_deadline": {
            "type": "string",
            "format": "date-time"
          },
          "grace_deadline": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "network": {
            "type": "string",
            "examples": [
              "eip155:8453"
            ]
          },
          "asset": {
            "type": "string",
            "examples": [
              "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
              "0x60a3E35Cc302bFA44Cb288Bc5a4F316Fdb1adb42"
            ]
          },
          "unlock_amount": {
            "type": "string",
            "description": "The one unlock amount, in the bound asset's atomic units.",
            "examples": [
              "3400000",
              "1700000",
              "2900000",
              "1450000"
            ]
          }
        }
      },
...
      "VerifyState": {
        "type": "object",
        "description": "The state of one chain. verdict, explanation, and response stay null until the unlock has been passed.",
        "properties": {
          "verify_id": {
            "type": "string",
            "format": "uuid",
            "description": "The durable handle for this chain. Persist it."
          },
          "work_id": {
            "type": "string",
            "format": "uuid",
            "description": "Alias of verify_id, as the service-windows extension names it."
          },
          "contract_version": {
            "type": "integer",
            "enum": [
              2,
              3
            ],
            "description": "3 for chains admitted under the three-gate contract. 2 for chains that finish under the old fixed USDC prices."
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "ready",
              "failed",
              "completed"
            ],
            "description": "processing: admission is settling or a human is working. ready: a human answered but the result is locked. failed: stop. completed: the unlock passed and the answer is available."
          },
          "human_status": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "accepted",
              "rejected",
              "refined",
              null
            ],
            "description": "The raw human action. Null until completed."
          },
          "verdict": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "true",
              "false",
              "refined",
              null
            ],
            "description": "The parseable verdict. true: the claim holds. false: it does not. refined: the human supplied a better version in explanation. Null until completed."
          },
          "explanation": {
            "type": [
              "string",
              "null"
            ],
            "description": "The human's free text when the verdict is refined. Null until completed."
          },
          "response": {
            "type": [
              "string",
              "null"
            ],
            "description": "The raw human response text. Null until completed."
          },
          "response_time_ms": {
            "type": [
              "integer",
              "null"
            ],
            "description": "How long the human took. Null until completed."
          },
          "wallet_address": {
            "type": [
              "string",
              "null"
            ],
            "description": "The requester wallet that owns this chain."
          },
          "funding": {
            "$ref": "#/components/schemas/Funding"
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "admitted_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the admission settlement was recorded and the chain entered the human queue. The windows start here. Null on the 202 of a paid admission."
          },
          "sla_deadline": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "admitted_at plus 60 minutes. An answer at or before it unlocks at the SLA price."
          },
          "grace_deadline": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "admitted_at plus 24 hours. An answer after sla_deadline and at or before it unlocks at the grace price."
          },
          "ready_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the human answered. It alone decides the unlock price."
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When unanswered work fails with human_timeout. On a contract 3 chain it equals grace_deadline, the last deadline, once the chain is admitted: 24 hours after admitted_at. Chains from before contract 3 expire 60 minutes after creation. It is null on the 202 of a paid admission, because the windows have not started yet; read it from the first poll."
          },
          "responded_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "unlocked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the unlock passed. Null in the unlock response itself; set on later polls."
          },
          "terms": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoundTerms"
              },
              {
                "type": "null"
              }
            ],
            "description": "The terms bound at admission. Null on chains from before contract 3."
          },
          "next_action": {
            "type": "string",
            "enum": [
              "poll",
              "unlock",
              "stop",
              "done"
            ],
            "description": "What the agent should do next. Follow this instead of inferring from status."
          },
          "poll_url": {
            "type": "string",
            "description": "Present when next_action is poll."
          },
          "retry_after_seconds": {
            "type": "integer",
            "description": "Present when next_action is poll. Honor it.",
            "examples": [
              15
            ]
          },
          "unlock_url": {
            "type": "string",
            "description": "Absolute unlock URL. Present when next_action is unlock.",
            "examples": [
              "https://verifi.cloud/verify-unlock?id=0f3f7d0a-6a1c-4a1e-9d7a-2f6b5b0f9c11"
            ]
          },
          "unlock": {
            "$ref": "#/components/schemas/UnlockAction"
          },
          "service_window": {
            "$ref": "#/components/schemas/ServiceWindow"
          },
          "failure": {
            "$ref": "#/components/schemas/FailureInfo"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong."
          },
          "detail": {
            "type": "string",
            "description": "What to do about it, when there is a useful next step."
          },
          "status": {
            "type": "string",
            "description": "The chain status, on lifecycle conflicts."
          }
        }
      }
    }
  }
}